refactor(agent/agentcontainers): Implement API service
