+++ agent/api.go
@@ -36,10 +36,15 @@ func (a *agent) apiHandler() http.Handler {
 		ignorePorts:   cpy,
 		cacheDuration: cacheDuration,
 	}
-	ch := agentcontainers.New(agentcontainers.WithLister(a.lister))
+
+	containerAPI := agentcontainers.NewAPI(
+		a.logger.Named("containers"),
+		agentcontainers.WithLister(a.lister),
+	)
+
 	promHandler := PrometheusMetricsHandler(a.prometheusRegistry, a.logger)
-	r.Get("/api/v0/containers", ch.List)
-	r.Post("/api/v0/containers/{id}/recreate", ch.Recreate)
+
+	r.Mount("/api/v0/containers", containerAPI.Routes())
 	r.Get("/api/v0/listening-ports", lp.handler)
 	r.Get("/api/v0/netcheck", a.HandleNetcheck)
 	r.Post("/api/v0/list-directory", a.HandleLS)
